// packages/plugin-calendar-ui/src/settings/types.ts
export type openingHour = {
  days: [string];
  start: string;
  end: string;
};

export type additionalField = {
  label: string;
  name: string;
  required: boolean;
  type: string;
};

export type SchedulePageMutationVariables = {
  _id?: string;
  name: string;
  slug: string;
  timezone: string;
  calendarIds: [string];
  event: {
    title: string;
    location: string;
    duration: number;
  };
  appearance?: {
    color: string;
    companyName?: string;
    logo?: string;
    submitText?: string;
    thankYouText?: string;
  };
  booking?: {
    openingHours?: openingHour[];
    additionalFields?: additionalField[];
    cancellationPolicy?: string;
    confirmationMethod?: string;
    minBookingNotice?: number;
    availableDaysInFuture?: number;
    minBuffer?: number;
    minCancellationNotice?: number;
  };
};

export type CreateSchedulePageMutationResponse = {
  createMutation: (params: {
    variables: SchedulePageMutationVariables;
    refetchQueries?: any[];
  }) => Promise<void>;
};

export type EditSchedulePageMutationResponse = {
  editMutation: (params: {
    variables: SchedulePageMutationVariables;
    refetchQueries?: any[];
  }) => Promise<void>;
};